+++ pkg/ipamapis/apis/fic/v1/types.go
@@ -38,10 +38,14 @@ type F5IPAMSpec struct {
 }
 
 type HostSpec struct {
-	Host      string `json:"host,omitempty"`
+	Host string `json:"host,omitempty"`
+
 	Key       string `json:"key,omitempty"`
-	CIDR      string `json:"cidr,omitempty"`
 	IPAMLabel string `json:"ipamLabel,omitempty"`
+
+	CIDR    string `json:"cidr,omitempty"`
+	NetView string `json:"netView,omitempty"`
+	DNSView string `json:"dnsView,omitempty"`
 }
 
 type F5IPAMStatus struct {
@@ -49,11 +53,15 @@ type F5IPAMStatus struct {
 }
 
 type IPSpec struct {
-	Host      string `json:"host,omitempty"`
+	IP   string `json:"ip,omitempty"`
+	Host string `json:"host,omitempty"`
+
 	Key       string `json:"key,omitempty"`
-	CIDR      string `json:"cidr,omitempty"`
 	IPAMLabel string `json:"ipamLabel,omitempty"`
-	IP        string `json:"ip,omitempty"`
+
+	CIDR    string `json:"cidr,omitempty"`
+	NetView string `json:"netView,omitempty"`
+	DNSView string `json:"dnsView,omitempty"`
 }
 
 // +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object